#1363 Move ShardingTransactionManager back to sharding-transaction-api.
