+++ sharding-transaction/sharding-transaction-api/src/main/java/io/shardingsphere/transaction/manager/ShardingTransactionManager.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.spi.transaction;
+package io.shardingsphere.transaction.manager;
 
 import io.shardingsphere.core.constant.transaction.TransactionType;
 import io.shardingsphere.core.event.transaction.ShardingTransactionEvent;
