+++ sharding-transaction/sharding-transaction-xa/src/main/java/io/shardingsphere/transaction/xa/handler/XAShardingTransactionHandler.java
@@ -19,7 +19,7 @@ package io.shardingsphere.transaction.xa.handler;
 
 import io.shardingsphere.core.constant.transaction.TransactionType;
 import io.shardingsphere.core.event.transaction.xa.XATransactionEvent;
-import io.shardingsphere.spi.transaction.ShardingTransactionManager;
+import io.shardingsphere.transaction.manager.ShardingTransactionManager;
 import io.shardingsphere.transaction.handler.ShardingTransactionHandlerAdapter;
 import io.shardingsphere.transaction.xa.manager.XATransactionManagerSPILoader;
